--- lib/pokotarou/additional_variables/def_variable.rb
@@ -1,3 +1,0 @@
-def const; AdditionalVariables.const end
-
-def let; AdditionalVariables.let end
