--- lib/pokotarou/arguments/def_args.rb
@@ -1,1 +1,0 @@
-def args; Arguments.args end
