+++ lib/pokotarou/registration_config_maker/import_option_setter.rb
@@ -1,9 +1,9 @@
-require "pokotarou/additional_methods.rb"
+require "pokotarou/additional_methods/main"
 module RegistrationConfigMaker
   class ImportOptionSetter
     class << self
       def set all_content
-        ::AdditionalMethods.import_from_yml(all_content[:"import'"])
+        ::AdditionalMethods::Main.import_from_yml(all_content[:"import'"])
         all_content.delete(:"import'")
       end
     end
